First pass at edge valuation
The main update is that I created a first real try at valuating a
player’s track choices. It could serve as a standard option in the
strategy. Combined with this was the proper creation of valuatecolors.m
which includes the values of the edges. Additionally, I began creating
a formal strategy set for valuateedges.m.

I combined points, pieces, and players into a structure called info.

I created plotgraph.m, which has the quick ability of showing the graph
layout with certain features highlighted.