Memory
Added a memory structure to allow for players to have a memory for
several reasons. Currently it is only being used to revaluate edges
once the current board has changed (someone laid a track). In the
future, players could remember something about others’ hands, etc.

Updated the probability of player laying down a track. I now use a
logistic with weights on the number of cards, edge value, and number of
players to determine the probability of putting a track.