Logistic function, track valuation fix
Created a general function called logistic which takes a variable
number of game considerations and uses the strategy to get a
probability of doing some action.

Fixed a edgevaluation problem. It was only looking for the best paths
between cities connected to the goals without looking through
intermediate clusters. I allowed the search to move without penalty
over edges already taken by the player, which fixed the problem.

Added points values for goals in plotgraph.