Various updates
Implemented a strategy for valuating edges’ weights using the points
earned, number of tracks to place, and number of turns required for
edges.

Fixed some issues with running out of cards.

Updated the logistic function to glm which can do all the same things
and more.

Updated the README!

Filled out the TODO to be more specific.