Genetic algorithm
Created two new functions for the genetic algorithm: population and
multiply. The first begins the genetic algorithm by randomizing
individuals’ features, having them play each other, and killing off the
loser. Then the winners multiply from pairs and then starts again.

Began work on a better algorithm for measuring the longest road. The
purpose of this is to make it easier for a player to estimate how
attainable longest road is, and also just to speed up the current
measurement.

Updated the strategy to remove all nested items. This makes it easier
for the genetic algorithm to combine parents.

Small fixes for the endgame and double color tracks.